fix: current profile could be inconsistent with *SessionHolder during store dehydration (#5394)
This commit addresses an issue with session resumption by updating the query keys for profile data to include a `isLogin` flag. This ensures that profile queries are correctly invalidated and refetched when the login status changes, preventing stale data issues.

The `createSelectors` helper was also updated to support custom selectors, which is used to prevent `useCurrentProfile` from returning an outdated `currentProfile` while the `useProfileStore` status is `Pending`.

This change also updates the `SetQueryDataForPosts` to set the profile data based on query key group that includes `isLogin` flag. This correctly caches the profile data for all possible `isLogin` states.